% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bkg_geocode.R, R/bkg_reverse.R
\encoding{UTF-8}
\name{bkg_geocode}
\alias{bkg_geocode}
\alias{bkg_reverse}
\title{BKG geocoding interface}
\usage{
bkg_geocode(
  .data,
  cols = 1L:4L,
  epsg = 3035,
  ...,
  join_with_original = FALSE,
  target_quality = NULL,
  verbose = TRUE
)

bkg_reverse(
  .data,
  epsg = 3035,
  target_quality = NULL,
  ...,
  join_with_original = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{.data}{For \code{bkg_reverse}, .data is expected to be an \code{sf}
data.frame containing geometries of a single geometry type. Geometry types
correspond to the allowed geometry types in
\code{\link[bkggeocoder]{bkg_reverse_single}}}

\item{cols}{\code{[numeric/character]}

Names or indices of the columns containing relevant geocoding information.
Must be of length 3 or 4. If a length-3 vector is passed, the first column is
interpreted as a single character string containing street and house number.
By default, interprets the first four columns as street, house number, zip
code and municipality (in this order).}

\item{epsg}{\code{[numeric/character]}

Numeric or character string containing an EPSG code for the requested CRS.}

\item{...}{Further arguments passed to
\code{\link[bkggeocoder]{bkg_geocode_single}} or
\code{\link[bkggeocoder]{bkg_reverse_single}}}

\item{join_with_original}{\code{[logical]}

Whether the input data should be joined with the output data. If \code{FALSE},
input data is discarded. Defaults to \code{TRUE}.}

\item{target_quality}{\code{[numeric]}

Targeted quality of the geocoding result. Only results are returned that
lie above this threshold. Note that unstructured}

\item{verbose}{\code{[logical]}

Whether to print informative messages and progress bars during
the geocoding process.}
}
\value{
\code{bkg_geocode} returns a nested list of class GeocodingResults
containing an \code{sf} dataframe of the geocoding results (\code{$geocoded})
and a dataframe with addresses with non-matched places (\code{$not_geocoded}).
Since the BKG geocoder does not need to subset the data using place matching,
the output does not contain dataframes on unmatched places like in
\code{\link{bkg_geocode_offline}}. The object also includes a call object.
Please note that original columns retrieve the suffix \code{"_input"}.
}
\description{
Geocoding of single or multiple addresses using the BKG geocoding WFS
interface (\code{wfs_geokodierung_bund}).
}
\examples{

\dontrun{
# dataset with addresses
address_data <- tibble::tribble(
 ~street, ~house_number, ~zip_code, ~place,
 "B2", "1", "68159", "Mannheim",
 "Unter Sachsenhausen", "6-8", "50667", "Köln"
)

bkg_geocode(data = address_data, epsg  = 4326)
}

}
