% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bkg_geocode_single_address.R
\name{bkg_geocode_single_address}
\alias{bkg_geocode_single_address}
\title{Geocoding of a single address to a geo-coordinate}
\usage{
bkg_geocode_single_address(
  street,
  house_number,
  zip_code,
  place,
  epsg = "3035"
)
}
\arguments{
\item{street}{Character string for the street name}

\item{house_number}{Character string for the house number}

\item{zip_code}{Character string for the zip code}

\item{place}{Character string for the place (i.e., municipality/city)}

\item{epsg}{Character string for the requested CSR}
}
\description{
Geocoding of a single address to a geo-coordinate through the BKG Geocoder
}
\examples{

\dontrun{
bkg_geocode_single_address(
  street       = "Unter Sachsenhausen",
  house_number = "6-8",
  zip_code     = 50667,
  place        = "Köln",
  epsg         = "3035"
)
}

}
