% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bkggeocoder-package.R
\docType{package}
\name{bkggeocoder-package}
\alias{bkggeocoder}
\alias{bkggeocoder-package}
\title{bkggeocoder: R Package for geocoding based on the data of the Federal Agency of Cartography and Geodesy (BKG) in Germany}
\description{
R Package for geocoding based on the data of the Federal Agency of Cartography and Geodesy (BKG) in Germany.
}
\author{
\strong{Maintainer}: Stefan Jünger \email{stefan.juenger@gesis.org}

}
\keyword{internal}
