% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.geocoding_results}
\alias{plot.geocoding_results}
\title{Get a simple results plot of the GeocodingResults class}
\usage{
\method{plot}{geocoding_results}(x, ...)
}
\arguments{
\item{x}{Object of class \code{GeocodingResults}}

\item{...}{Further arguments passed on to \code{\link[graphics]{hist}}}
}
\description{
Get a simple results plot of the GeocodingResults class
}
