% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.GeocodingResults}
\alias{plot.GeocodingResults}
\title{Get a simple results plot of the GeocodingResults class}
\usage{
\method{plot}{GeocodingResults}(x, ...)
}
\arguments{
\item{x}{R object of class 'GeocodingResults'}

\item{...}{Further arguments passed on to \code{\link[graphics]{hist}}}
}
\description{
Get a simple results plot of the GeocodingResults class
}
