% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.geocoding_results}
\alias{print.geocoding_results}
\title{Print the output of the GeocodingResults class}
\usage{
\method{print}{geocoding_results}(
  x,
  which = c("geocoded", "not_geocoded", "not_place_matched", "unmatched_places"),
  ...
)
}
\arguments{
\item{x}{Object of class \code{GeocodingResults}}

\item{which}{character string; choose which type of the results should be
displayed}

\item{...}{Further arguments passed on to
\code{\link[base:print.data.frame]{base::print.data.frame()}}
or \code{\link[sf:sf]{sf::print.sf()}}}
}
\description{
Print the output of the GeocodingResults class
}
