% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.GeocodingResults}
\alias{print.GeocodingResults}
\title{Print the output of the GeocodingResults class}
\usage{
\method{print}{GeocodingResults}(x, which = c("successful", "na", "unmatched_places"), ...)
}
\arguments{
\item{x}{R object of class 'GeocodingResults'}

\item{which}{character string; choose which type of the results should be
displayed}
}
\description{
Print the output of the GeocodingResults class
}
