% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.geocoding_results}
\alias{summary.geocoding_results}
\title{Get a summary of the GeocodingResults class}
\usage{
\method{summary}{geocoding_results}(object, ...)
}
\arguments{
\item{object}{Object of class \code{GeocodingResults}}

\item{...}{Ignored.}
}
\description{
Get a summary of the GeocodingResults class
}
