% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bkg_export_geocodes.R
\name{bkg_export_geocodes}
\alias{bkg_export_geocodes}
\title{Export geocoding results}
\usage{
bkg_export_geocodes(
  data,
  which = c("all", "geocoded", "not_geocoded", "not_place_matched", "unmatched_places"),
  file,
  overwrite = TRUE,
  ...
)
}
\arguments{
\item{data}{R object of class GeocodingResults; the geocoded addresses}

\item{which}{Character string; what kind of results would you like to export
(default is the successful and na ones appended together)}

\item{file}{Path to the output file. The file type is guessed by
\code{\link[sf]{st_write}} and must be supported by
\code{\link[sf]{st_drivers}}.}

\item{overwrite}{Logical; define whether exported file get overwritten
(default is true)}

\item{...}{Further arguments passed to \code{\link[utils]{write.table}} or
\code{\link[sf]{st_write}}.}
}
\value{
\code{file}, invisibly.
}
\description{
Export the output of \code{\link[bkggeocoder]{bkg_geocode_offline}} and
\code{\link[bkggeocoder]{bkg_geocode}}.
}
