% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bkg_geocode.R, R/bkg_geocode_single_address.R
\name{bkg_geocode}
\alias{bkg_geocode}
\alias{bkg_geocode_single_address}
\title{BKG geocoding interface}
\usage{
bkg_geocode(
  data,
  cols = 1L:4L,
  epsg = 3035,
  target_quality = 0.9,
  verbose = TRUE
)

bkg_geocode_single_address(
  street,
  house_number,
  zip_code,
  place,
  epsg = 3035,
  max_features = 1
)
}
\arguments{
\item{epsg}{Numeric or character string containing an EPSG code for the
requested CRS.}

\item{target_quality}{Numeric; targeted quality. All geocoded addresses with
scores below this value will be dropped in the output.}

\item{street}{Character string for the street name}

\item{house_number}{Character string for the house number}

\item{zip_code}{Character string for the zip code}

\item{place}{Character string for the place (i.e., municipality/city)}

\item{max_features}{Numeric or character specifying the maximum amount of
results to return. If \code{> 1}, returns matches in descending order by
BKG score.}
}
\value{
Returns a tidy simple features data frame of the original data object
and the requested information from the geocoding service. Please note that
all columns are converted to character strings.
}
\description{
Geocoding of single or multiple addresses using the BKG geocoding WFS
interface (\code{wfs_geokodierung_bund}).
}
\examples{

\dontrun{
# dataset with addresses
address_data <- tibble::tribble(
 ~street, ~house_number, ~zip_code, ~place,
 "B2", "1", "68159", "Mannheim",
 "Unter Sachsenhausen", "6-8", "50667", "Köln"
)

bkg_geocode(data = address_data, epsg  = 4326)
}

\dontrun{
# single address
bkg_geocode_single_address(
  street       = "Unter Sachsenhausen",
  house_number = "6-8",
  zip_code     = 50667,
  place        = "Köln",
  epsg         = 3035
)
}

}
