% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bkg_geocode.R, R/bkg_geocode_single_address.R
\name{bkg_geocode}
\alias{bkg_geocode}
\alias{bkg_geocode_single_address}
\title{BKG geocoding interface}
\usage{
bkg_geocode(
  data,
  cols = 1L:4L,
  epsg = 3035,
  target_quality = 0.9,
  verbose = TRUE
)

bkg_geocode_single_address(
  street,
  house_number,
  zip_code,
  place,
  epsg = 3035,
  max_features = 1
)
}
\arguments{
\item{data}{Dataframe containing address data. The dataframe must contain
at least four columns carrying the street name, house number, zip code and
municipality name. The corresponding column names or indices can be specified
using the \code{cols} argument.}

\item{cols}{Numeric or character of length 4; names or indices of the columns
containing relevant geocoding information. By default, interprets the
first four columns as street, house number, zip code and municipality (in
this order).}

\item{epsg}{Numeric or character string containing an EPSG code for the
requested CRS.}

\item{target_quality}{Numeric; targeted quality of second record linkage
round (see details). Corresponds to the threshold value of
\code{\link[reclin2]{jaro_winkler}} and \code{\link[reclin2]{select_greedy}}.}

\item{verbose}{Whether to print informative messages and progress bars during
the geocoding process.}

\item{street}{Character string for the street name}

\item{house_number}{Character string for the house number}

\item{zip_code}{Character string for the zip code}

\item{place}{Character string for the place (i.e., municipality/city)}

\item{max_features}{Numeric or character specifying the maximum amount of
results to return. If \code{> 1}, returns matches in descending order by
BKG score.}
}
\value{
Returns a tidy simple features data frame of the original data object
and the requested information from the geocoding service. Please note that
all columns are converted to character strings.
}
\description{
Geocoding of single or multiple addresses using the BKG geocoding WFS
interface (\code{wfs_geokodierung_bund}).
}
\examples{

\dontrun{
# dataset with addresses
address_data <- tibble::tribble(
 ~street, ~house_number, ~zip_code, ~place,
 "B2", "1", "68159", "Mannheim",
 "Unter Sachsenhausen", "6-8", "50667", "Köln"
)

bkg_geocode(data = address_data, epsg  = 4326)
}

\dontrun{
# single address
bkg_geocode_single_address(
  street       = "Unter Sachsenhausen",
  house_number = "6-8",
  zip_code     = 50667,
  place        = "Köln",
  epsg         = 3035
)
}

}
