% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bkg_geocode.R
\encoding{UTF-8}
\name{bkg_geocode}
\alias{bkg_geocode}
\title{BKG geocoding interface}
\usage{
bkg_geocode(
  data,
  cols = 1L:4L,
  epsg = 3035,
  join_with_original = FALSE,
  target_quality = 0.9,
  verbose = TRUE
)
}
\arguments{
\item{data}{\code{[data.frame]}

Dataframe containing address data. The dataframe must contain columns carrying
the street name, house number, zip code and municipality name. The
corresponding column names or indices can be specified using the \code{cols}
argument.}

\item{cols}{\code{[numeric/character]}

Names or indices of the columns containing relevant geocoding information.
Must be of length 3 or 4. If a length-3 vector is passed, the first column is
interpreted as a single character string containing street and house number.
By default, interprets the first four columns as street, house number, zip
code and municipality (in this order).}

\item{epsg}{Numeric or character string containing an EPSG code for the
requested CRS.}

\item{join_with_original}{\code{[logical]}

Whether the input data should be joined with the output data. If \code{FALSE},
input data is discarded. Defaults to \code{TRUE}.}

\item{target_quality}{Numeric; targeted quality. All geocoded addresses with
scores below this value will be dropped in the output.}

\item{verbose}{\code{[logical]}

Whether to print informative messages and progress bars during
the geocoding process.}
}
\value{
\code{bkg_geocode}r eturns a nested list of class GeocodingResults
containing an \code{sf} dataframe of the geocoding results (\code{$geocoded})
and a dataframe with addresses with non-matched places (\code{$not_geocoded}).
Since the BKG geocoder does not need to subset the data using place matching,
the output does not contain dataframes on unmatched places like in
\code{\link{bkg_geocode_offline}}. The object also includes a call object and
descriptive summary statistics. Please note that original columns
retrieve the suffix \code{"_input"}.
}
\description{
Geocoding of single or multiple addresses using the BKG geocoding WFS
interface (\code{wfs_geokodierung_bund}).
}
\examples{

\dontrun{
# dataset with addresses
address_data <- tibble::tribble(
 ~street, ~house_number, ~zip_code, ~place,
 "B2", "1", "68159", "Mannheim",
 "Unter Sachsenhausen", "6-8", "50667", "Köln"
)

bkg_geocode(data = address_data, epsg  = 4326)
}

}
