% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bkg_geocode.R
\name{bkg_geocode}
\alias{bkg_geocode}
\title{Geocoding of a multiple addresses to a geo-coordinate}
\usage{
bkg_geocode(data = NULL, cols = 1L:4L, epsg = "3035", verbose = TRUE)
}
\arguments{
\item{data}{Dataframe containing address data. The dataframe must contain
at least four columns carrying the street name, house number, zip code and
municipality name. The corresponding column names or indices can be specified
using the \code{cols} argument.}

\item{cols}{Numeric or character of length 4; names or indices of the columns
containing relevant geocoding information. By default, interprets the
first four columns as street, house number, zip code and municipality (in
this order).}

\item{epsg}{Character string for the requested CSR}

\item{verbose}{Whether to print informative messages and progress bars during
the geocoding process.}
}
\value{
Returns a tidy simple features data frame of the original data object
and the requested information from the geocoding service. Please note that
all columns are converted to character strings.
}
\description{
Geocoding of a multiple addresses to a geo-coordinate through the BKG Geocoder
}
\examples{

\dontrun{
# single address
bkg_geocode_single_address(
  street       = "Unter Sachsenhausen",
  house_number = "6-8",
  zip_code     = 50667,
  place        = "Köln",
  epsg         = "3035"
)

# dataset with addresses
address_data <-
  tibble::tribble(
    ~street, ~house_number, ~zip_code, ~place,
    "B2", "1", "68159", "Mannheim",
    "Unter Sachsenhausen", "6-8", "50667", "Köln"
  )

address_data_geocoded <-
  bkg_geocode(
    data         = address_data,
    street       = "street",
    house_number = "house_number",
    zip_code     = "zip_code",
    place        = "place",
    epsg         = "3035"
  )
}

}
