% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bkg_geocode.R
\name{bkg_geocode}
\alias{bkg_geocode}
\title{Geocoding of a multiple addresses to a geo-coordinate}
\usage{
bkg_geocode(
  data = NULL,
  street = "street",
  house_number = "house_number",
  zip_code = "zip_code",
  place = "place",
  epsg = "3035"
)
}
\arguments{
\item{data}{Data object that contains the columns for the parameters defined
next}

\item{street}{Character string for the street name}

\item{house_number}{Character string for the house number}

\item{zip_code}{Character string for the zip code}

\item{place}{Character string for the place (i.e., municipality/city)}

\item{epsg}{Character string for the requested CSR}
}
\value{
Returns a tidy simple features data frame of the original data object
and the requested information from the geocoding service. Please note that
all columns are converted to character strings.
}
\description{
Geocoding of a multiple addresses to a geo-coordinate through the BKG Geocoder
}
\examples{

\dontrun{
# single address
bkg_geocode_single_address(
  street       = "Unter Sachsenhausen",
  house_number = "6-8",
  zip_code     = 50667,
  place        = "Köln",
  epsg         = "3035"
)

# dataset with addresses
address_data <-
  tibble::tribble(
    ~street, ~house_number, ~zip_code, ~place,
    "B2", "1", "68159", "Mannheim",
    "Unter Sachsenhausen", "6-8", "50667", "Köln"
  )

address_data_geocoded <-
  bkg_geocode(
    data         = address_data,
    street       = "street",
    house_number = "house_number",
    zip_code     = "zip_code",
    place        = "place",
    epsg         = "3035"
  )
}

}
