% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bkg_geocode_offline.R
\encoding{UTF-8}
\name{bkg_geocode_offline}
\alias{bkg_geocode_offline}
\title{Geocoding of a multiple addresses (BKG offline version)}
\usage{
bkg_geocode_offline(
  data,
  cols = 1L:4L,
  data_from_server = FALSE,
  data_path = "../bkgdata/",
  credentials_path = "../bkgcredentials/",
  join_with_original = TRUE,
  crs = 3035L,
  place_match_quality = 0.9,
  place_match_opts = list(),
  target_quality = 0.9,
  target_opts = list(),
  verbose = TRUE,
  force_decrypt = FALSE
)
}
\arguments{
\item{data}{Dataframe containing address data. The dataframe must contain
at least four columns carrying the street name, house number, zip code and
municipality name. The corresponding column names or indices can be specified
using the \code{cols} argument.}

\item{cols}{Numeric or character of length 4; names or indices of the columns
containing relevant geocoding information. By default, interprets the
first four columns as street, house number, zip code and municipality (in
this order).}

\item{data_from_server}{Logical; shall the address data be downloaded from
GESIS internal server? Requires access to the GESIS net (default is
\code{FALSE})}

\item{data_path}{Path to the address data provided by Stefan Jünger. Ignored
if \code{data_from_server = TRUE}.}

\item{credentials_path}{Path to credentials package provided by Stefan Jünger}

\item{join_with_original}{Logical; should the output be joined with the
input data?}

\item{crs}{Any kind of object that can be parsed by \code{\link[sf]{st_crs}}
that the output data should be transformed to (e.g. EPSG code, WKT/PROJ4
character string or object of class \code{crs}). Defaults to EPSG:3035.}

\item{place_match_quality}{Numeric; targeted quality of first record linkage
round (see details). Corresponds to the threshold value of
\code{\link[reclin2]{jaro_winkler}}.}

\item{target_quality}{Numeric; targeted quality of second record linkage
round (see details). Corresponds to the threshold value of
\code{\link[reclin2]{jaro_winkler}} and \code{\link[reclin2]{select_greedy}}.}

\item{verbose}{Whether to print informative messages and progress bars during
the geocoding process.}

\item{force_decrypt}{Whether to force the function to read in the encrypted
files. This can be useful, if the BKG data should not be stored locally for
too long or if the cached data are corrupt or outdated.}
}
\value{
Returns a nested list of class GeocodingResult containing an
\code{sf} dataframe of the geocoding results (\code{$geocoded_data}) as well
as a dataframe with non-matched place names (\code{$unmatched_places}),
addresses with non-matched places (\code{$non_geocoded_data}) and non-matched
addresses (\code{$geocoded_data_na}). The object also includes a call
object and descriptive summary statistics. Please note that original columns
retrieved the suffix \code{"_input"}.
}
\description{
Geocoding of a multiple addresses using record linkage and an
address/coordinate database (provided by the BKG)
}
\details{
The function first matches the zip code and place information from
the data against the official names in the address/geocoordinate database
(first round of record linkage). This is done to filter out address datasets
that are not needed and lower the data size. You can play with the quality by
adjusting the \code{place_match_quality} parameter. In a second step, the
input addresses together with the matched results are then again matched
against the addresses in the address/geocoordinate database (second round of
record linkage). Again, you can play with the quality by adjusting
the \code{target_quality} parameter.

Record linkage is employed using string distance metrics from
\code{\link[stringdist]{stringdist}}. By default, scores for both place
matching and geocoding are calculated using the standard Jaro distance. This
metric can be adjusted by passing a list of options that is then passed on
to \code{\link[stringdist]{stringdist}}. This can include the method of
choice \code{method} as well as further method-specific parameters like
the size of the q-gram (\code{q}), the Jaro-Winkler prefix factor (\code{p}),
and the Winkler boost threshold (\code{bt}). To derive a distance index that
falls between 0 and 1, where 0 denotes a complete dissimilarity between both
address strings and 1 denotes a complete similarity, certain edit and q-gram
distance metrics are devided by their maximum possible value as explained in
van der Loo (2014).
For more details on the method
choice, refer to the \code{\link[stringdist]{stringdist-metrics}}
documentation from the \code{stringdist} package.

The overall quality of the geocoding can be evaluated by looking at the
values of the column \code{score} (ranging from 0 to 1), which is based on
the second round of record linkage. In general, for both rounds of record
linkage, a score of above 0.9 can be considered a good match. If the score
falls below 0.8, the result should be thoroughly scrutinized.

Address data loading works using a temporary cache. Before trying to request
and decrypt encrypted data chunks, the function will look for a file named
\code{bkg_data_cache.rda} in \code{tempdir()}. If found, data decrypting will
be skipped which can shorten the processing time significantly for large
input datasets with a high number of different places. If needed, this
behavior can be suppressed by setting \code{force_decrypt = TRUE}.
}
\references{
van der Loo, M. P. J. (2014). The stringdist Package for
Approximate String Matching. The R Journal, 6(1), 111–122. https://doi.org/10.32614/RJ-2014-011
}
