% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bkg_geocode_offline.R
\name{bkg_geocode_offline}
\alias{bkg_geocode_offline}
\title{Geocoding of a multiple addresses (BKG offline version)}
\usage{
bkg_geocode_offline(
  data,
  id_variable = "id",
  street = "street",
  house_number = "house_number",
  zip_code = "zip_code",
  place = "place",
  data_path = "../bkgdata/",
  credentials_path = "../bkgcredentials/",
  join_with_original = TRUE,
  epsg = NULL,
  place_match_quality = 0.5,
  target_quality = 0.5,
  echo = TRUE
)
}
\arguments{
\item{data}{Data object that contains the columns for the parameters defined
next}

\item{street}{Character string for the street name}

\item{house_number}{Character string for the house number}

\item{zip_code}{Character string for the zip code}

\item{place}{Character string for the place (i.e., municipality/city)}

\item{credentials_path}{path to credentials package provided by Stefan Jünger}

\item{join_with_original}{logical; join with original data or only output
results}

\item{epsg}{Character string for the requested CSR}

\item{place_match_quality}{numeric; targeted quality of first record linkage
round (see details)}

\item{target_quality}{numeric; targeted quality of second record linkage
round (see details)}
}
\value{
Returns a tidy simple features data frame of the original data object
and the requested information from the geocoding. Please note that
original columns retrieved the suffix "_input".
}
\description{
Geocoding of a multiple addresses using record linkage and an
address/coordinate database (provided by the BKG)
}
\details{
The function first matches the zip code and place information from
the data against the official names in the address/geocoordinate database
(first round of record linkage). You can play with the quality by adjusting
the \code{place_match_quality} parameter. In a second step, the input
addresses together with the matched results are then again matched against
the addresses in the address/geocoordinate database (second round of record
linkage). Again, you can play with the quality by adjusting
the \code{target_quality} parameter. The overall quality of
the geocoding can be evaluated by looking at the values of the column
\code{score} (ranging from 0 to 1), which is based on the second round of
record linkage.¸
}
