% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bkg_geocode_offline.R
\encoding{UTF-8}
\name{bkg_geocode_offline}
\alias{bkg_geocode_offline}
\title{Geocoding of a multiple addresses (BKG offline version)}
\usage{
bkg_geocode_offline(
  data,
  cols = 1L:4L,
  data_from_server = FALSE,
  data_path = "../bkgdata/",
  credentials_path = "../bkgcredentials/",
  join_with_original = TRUE,
  crs = 3035,
  place_match_quality = 0.5,
  target_quality = 0.5,
  verbose = TRUE
)
}
\arguments{
\item{data}{Dataframe containing address data. The dataframe must contain
at least four columns carrying the street name, house number, zip code and
municipality name. The corresponding column names or indices can be specified
using the \code{cols} argument.}

\item{cols}{Numeric or character of length 4; names or indices of the columns
containing relevant geocoding the input data. By default, interprets the
first four columns as street, house number, zip code and municipality (in
this order).}

\item{data_from_server}{Logical; shall the address data be downloaded from
GESIS internal server? Requires access to the GESIS net (default is
\code{FALSE})}

\item{data_path}{Path to the address data provided by Stefan Jünger}

\item{credentials_path}{Path to credentials package provided by Stefan Jünger}

\item{join_with_original}{Logical; should the output be joined with the
input data?}

\item{crs}{EPSG code, WKT/PROJ4 character string or object of class \code{crs}
that the output should be transformed to. Defaults to EPSG:3035.}

\item{place_match_quality}{Numeric; targeted quality of first record linkage
round (see details). Corresponds to the threshold value of
\code{\link[reclin2]{jaro_winkler}}.}

\item{target_quality}{Numeric; targeted quality of second record linkage
round (see details). Corresponds to the threshold value of
\code{\link[reclin2]{jaro_winkler}} and \code{\link[reclin2]{select_greedy}}.}

\item{verbose}{Whether to print informative messages and progress bars during
the geocoding process.}
}
\value{
Returns a nested list of class GeocodingResult containing an
\code{sf} dataframe of the geocoding results (\code{$geocoded_data}) as well
as a dataframe with non-matched place names (\code{$unmatched_places}),
non-matched addresses (\code{$non_geocoded_data}) and matches with unresolved
regional key (\code{$geocoded_data_na}). The object also includes a call
object and descriptive summary statistics. Please note that original columns
retrieved the suffix \code{"_input"}.
}
\description{
Geocoding of a multiple addresses using record linkage and an
address/coordinate database (provided by the BKG)
}
\details{
The function first matches the zip code and place information from
the data against the official names in the address/geocoordinate database
(first round of record linkage). This is done to filter out address datasets
that are not needed and lower the data size. You can play with the quality by
adjusting the \code{place_match_quality} parameter. In a second step, the
input addresses together with the matched results are then again matched
against the addresses in the address/geocoordinate database (second round of
record linkage). Again, you can play with the quality by adjusting
the \code{target_quality} parameter. 
The overall quality of the geocoding can be evaluated by looking at the
values of the column \code{score} (ranging from 0 to 1), which is based on
the second round of record linkage.
}
