% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spt_create_inspire_ids.R
\name{spt_create_inspire_ids}
\alias{spt_create_inspire_ids}
\title{Create 1km and 100m INSPIRE IDs}
\usage{
spt_create_inspire_ids(
  data,
  type = c("1km", "100m"),
  column_name = "Gitter_ID_",
  combine = FALSE
)
}
\arguments{
\item{data}{Object of class ```sf``` containing point geometries}

\item{type}{Character string for the requested ID type}
}
\value{
tibble
}
\description{
Create 1 km² and 100m X 100m INSPIRE IDs from coordinates
}
