% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.GeocodingResults}
\alias{summary.GeocodingResults}
\title{Get a summary of the GeocodingResults class}
\usage{
\method{summary}{GeocodingResults}(x, ...)
}
\arguments{
\item{x}{R object of class 'GeocodingResults'}
}
\description{
Get a summary of the GeocodingResults class
}
